Split Definitions and move files in a sub folder
